src/execution/requirements.txt
# Execution Layer Dependencies
# 链上交易执行依赖

# Web3 交互
web3>=6.0.0
eth-account>=0.8.0

# HTTP 客户端
aiohttp>=3.8.0
requests>=2.28.0

# 环境变量
python-dotenv>=1.0.0

# Redis
redis>=5.0.0

# 日志
colorlog>=6.7.0

# YAML 配置
pyyaml>=6.0

# 其他


